+++ tests/bluetooth/bsim_bt/bsim_test_audio/src/csis_test.c
@@ -5,7 +5,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 #ifdef CONFIG_BT_CSIS
-#include "../../../../../subsys/bluetooth/audio/csis.h"
+#include <bluetooth/audio/csis.h>
 
 #include "common.h"
